+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 v1.1.1
+- Modified Fortigate authentication to use Authorization Bearer Token header rather than query string
 - Logging improvements
 - Upgraded README to use doctool